rename sections to components
